+++ src/components/FloatingDropdown/FloatingDropdown.visual.tsx
@@ -17,6 +17,8 @@ class FloatingDropdownVisual extends React.Component<any> {
   }
 }
 
+const veryLongText = 'Some very very very very very very very very long text';
+
 const extendedProps: Partial<FloatingDropdownProps> = {
   forceContentElementVisibility: true,
 };
@@ -31,17 +33,27 @@ visualize('FloatingDropdown', () => {
 
     snap(
       'expanded',
-      // tslint:disable-next-line:jsx-wrap-multiline
       <FloatingDropdownVisual
         {...getFloatingDropdownTestProps({ ...extendedProps })}
       />,
+    );
+
+    snap(
+      'ellipsis',
+      <div style={{ width: 200 }}>
+        <FloatingDropdownVisual
+          {...getFloatingDropdownTestProps({
+            options: [{ id: '1', value: veryLongText, isSelectable: true }],
+            value: '1',
+          })}
+        />
+      </div>,
     );
   });
 
   story('preselected', () => {
     snap(
       'default',
-      // tslint:disable-next-line:jsx-wrap-multiline
       <FloatingDropdownVisual
         {...getFloatingDropdownTestProps({ ...preselectedProps })}
       />,
@@ -49,7 +61,6 @@ visualize('FloatingDropdown', () => {
 
     snap(
       'expanded',
-      // tslint:disable-next-line:jsx-wrap-multiline
       <FloatingDropdownVisual
         {...getFloatingDropdownTestProps({
           ...preselectedProps,
@@ -62,17 +73,31 @@ visualize('FloatingDropdown', () => {
   story('disabled', () => {
     snap(
       'default',
-      // tslint:disable-next-line:jsx-wrap-multiline
       <FloatingDropdownVisual
         {...getFloatingDropdownTestProps({ disabled: true })}
       />,
     );
   });
 
+  story('displayBlock', () => {
+    snap(
+      'default',
+      <FloatingDropdownVisual
+        {...getFloatingDropdownTestProps({ displayBlock: true })}
+      />,
+    );
+
+    snap(
+      'mobile',
+      <TPAComponentsProvider value={{ mobile: true }}>
+        <FloatingDropdownVisual {...getFloatingDropdownTestProps({ displayBlock: true })} />
+      </TPAComponentsProvider>,
+    );
+  });
+
   story('native', () => {
     snap(
       'default',
-      // tslint:disable-next-line:jsx-wrap-multiline
       <TPAComponentsProvider value={{ mobile: true }}>
         <FloatingDropdownVisual {...getFloatingDropdownTestProps()} />
       </TPAComponentsProvider>,