+++ src/components/FloatingDropdown/docs/examples.ts
@@ -99,3 +99,50 @@ export const preselected = `
   />
 </ExampleWithContextProps>
 </>`;
+
+export const ellipsis = `
+<div style={{width: 200}}>
+  <FloatingDropdown
+    options={[
+      {id: '1', value: 'Very very very very very very very very very long text', isSelectable: true},
+    ]}
+    label="Sort by:"
+    placeholder="Very very very very very very very very very long placeholder"
+    value="1"
+  />
+</div>
+`;
+
+export const displayBlock = `
+<>
+<h2>Desktop Example</h2>
+<FloatingDropdown
+    options={[
+      {id: '1', value: 'Most recent', isSelectable: true},
+      {id: '2', value: 'Most viewed', isSelectable: true},
+      {id: '3', value: 'Most liked', isSelectable: true},
+      {id: '4', value: 'Most comments', isSelectable: true},
+      {id: '5', value: 'Most recent posts in the whole galaxy', isSelectable: true},
+    ]}
+    label="Sort by:"
+    placeholder="Something"
+    value="1"
+    displayBlock
+/>
+<h2>Native Example</h2>
+<ExampleWithContextProps mobile={true}>
+  <FloatingDropdown
+    options={[
+      {id: '1', value: 'Most recent', isSelectable: true},
+      {id: '2', value: 'Most viewed', isSelectable: true},
+      {id: '3', value: 'Most liked', isSelectable: true},
+      {id: '4', value: 'Most comments', isSelectable: true},
+      {id: '5', value: 'Most recent posts in the whole galaxy', isSelectable: true},
+    ]}
+    label="Sort by:"
+    placeholder="Something"
+    value="1"
+    displayBlock
+  />
+</ExampleWithContextProps>
+</>`;