+++ src/components/RadioButton/RadioButton.e2e.ts
@@ -1,5 +1,5 @@
 import * as eyes from 'eyes.it';
-import { browser } from 'protractor';
+import { browser, Key } from 'protractor';
 import {
   createStoryUrl,
   waitForVisibilityOf,
@@ -27,6 +27,8 @@ describe('RadioButton', () => {
       'Cannot find radioButton',
     );
     await driver.clickInput();
+    await browser.actions().sendKeys(Key.ARROW_DOWN, Key.ARROW_UP);
+    expect(await driver.isFocused()).toBeTruthy();
   });
 
   eyes.it('should show the correct design on focus theme default', async () => {
@@ -36,5 +38,7 @@ describe('RadioButton', () => {
       'Cannot find radioButton',
     );
     await driver.clickInput();
+    await browser.actions().sendKeys(Key.ARROW_DOWN, Key.ARROW_UP);
+    expect(await driver.isFocused()).toBeTruthy();
   });
 });